stt/requirements.txt
pyhocon
ffmpeg-python
sanic
jiwer
regex
